[reorganization]:Move locality update tool logic to its own app
Fixes #7342
