+++ specifyweb/backend/locality_update_tool/tests/test_parse_locality_set_foreground.py
@@ -1,6 +1,6 @@
 from unittest.mock import patch
 from specifyweb.specify.tests.test_api import ApiTests
-from specifyweb.specify.update_locality import ParseError, ParsedRow
+from specifyweb.backend.locality_update_tool.update_locality import ParseError, ParsedRow
 from specifyweb.specify.views import parse_locality_set_foreground
