+++ specifyweb/specify/tests/test_update_locality/test_merge_parse_results.py
@@ -1,7 +1,7 @@
 from specifyweb.specify.tests.test_update_locality.test_update_locality_context import (
     TestUpdateLocalityContext,
 )
-from specifyweb.specify.update_locality import (
+from specifyweb.backend.locality_update_tool.update_locality import (
     ParseError,
     ParseSuccess,
     merge_parse_results,
