+++ specifyweb/specify/tests/test_update_locality/test_resolve_localityupdate_result.py
@@ -5,7 +5,7 @@ from uuid import uuid4
 from specifyweb.specify.tests.test_update_locality.test_update_locality_context import (
     TestUpdateLocalityContext,
 )
-from specifyweb.specify.update_locality import (
+from specifyweb.backend.locality_update_tool.update_locality import (
     LocalityUpdateStatus,
     resolve_localityupdate_result,
 )
