+++ specifyweb/specify/tests/test_update_locality/test_upload_locality_set.py
@@ -1,7 +1,7 @@
 from specifyweb.specify.tests.test_update_locality.test_update_locality_context import (
     TestUpdateLocalityContext,
 )
-from specifyweb.specify.update_locality import upload_locality_set
+from specifyweb.backend.locality_update_tool.update_locality import upload_locality_set
 
 
 class TestUploadLocalitySet(TestUpdateLocalityContext):
