+++ specifyweb/specify/urls.py
@@ -57,14 +57,6 @@ urlpatterns = [
         path('repair/', tree_views.repair_tree),
     ])),
 
-    # locality set import endpoints
-    path('localityset/', include([
-        path('parse/', views.parse_locality_set),
-        path('import/', views.upload_locality_set),
-        re_path(r'^status/(?P<taskid>[0-9a-fA-F-]+)/$', views.localityupdate_status),
-        re_path(r'^abort/(?P<taskid>[0-9a-fA-F-]+)/$', views.abort_localityupdate_task),
-    ])),
-
     # generates Sp6 master key
     path('master_key/', master_key.master_key),
