+++ specifyweb/urls.py
@@ -20,6 +20,7 @@ from .specify.views import images, properties
 from specifyweb.backend.stored_queries import urls as query_urls
 from specifyweb.backend.workbench import urls as wb_urls
 from specifyweb.backend.stats import urls as stat_urls
+from specifyweb.backend.locality_update_tool import urls as locality_update_tool_urls
 
 # print(get_resolver().reverse_dict.keys()) # Use for debugging urls
 
@@ -61,5 +62,6 @@ urlpatterns = [
     path('export/', include(export_urls)), # permissions added
     path('permissions/', include(permissions_urls)), # permissions added
     # url(r'^testcontext/', include()),
-    path('stats/', include(stat_urls))
+    path('stats/', include(stat_urls)),
+    path('locality_update_tool/', include(locality_update_tool_urls))
 ]
